Fixed issue placing markers
Formerly, markers would not appear on my device because of a race condition. The method to place a marker was static, and it was being called potentially before the static variable mMap was instatiated within OnCreateActivity.

To solve the issue, I added a static marker variable that is set by findBuilding(). Now, the marker is placed when the activity is created. There is only one marker at a time, but the marker is not overwritten when the user changes views.